refactor: move useImportRestrictions into noRestrictedImports
