+++ .changeset/the_rule_useimportrestrictions_has_been_moved_into_norestrictedimports.md
@@ -1,0 +1,26 @@
+---
+cli: major
+---
+
+# The rule `useImportRestrictions` has been removed and replaced with an option in `noRestrictedImports`
+
+To avoid confusion between the two similarly named rules, `useImportRestrictions` has been removed
+and replaced with the option `restrictPackagePrivate` in the `noRestrictedImports` rule:
+
+```jsonc
+// biome.json
+{
+  "linter": {
+    "rules": {
+      "nursury": {
+        "noRestrictedImports": {
+          "level": "error",
+          "options": {
+            "restrictPackagePrivate": "all"
+          }
+        }
+      }
+    }
+  }
+}
+```
