+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -474,16 +474,12 @@ pub(crate) fn migrate_eslint_any_rule(
             rule.set_level(rule.level().max(rule_severity.into()));
         }
         "import-access/eslint-plugin-import-access" => {
-            if !options.include_inspired {
-                results.has_inspired_rules = true;
-                return false;
-            }
             if !options.include_nursery {
                 return false;
             }
             let group = rules.nursery.get_or_insert_with(Default::default);
             let rule = group
-                .use_import_restrictions
+                .no_restricted_imports
                 .get_or_insert(Default::default());
             rule.set_level(rule.level().max(rule_severity.into()));
         }
