+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -36,7 +36,6 @@ pub mod use_explicit_type;
 pub mod use_google_font_display;
 pub mod use_google_font_preconnect;
 pub mod use_guard_for_in;
-pub mod use_import_restrictions;
 pub mod use_sorted_classes;
 pub mod use_strict_mode;
 pub mod use_trim_start_end;
@@ -80,7 +79,6 @@ declare_lint_group! {
             self :: use_google_font_display :: UseGoogleFontDisplay ,
             self :: use_google_font_preconnect :: UseGoogleFontPreconnect ,
             self :: use_guard_for_in :: UseGuardForIn ,
-            self :: use_import_restrictions :: UseImportRestrictions ,
             self :: use_sorted_classes :: UseSortedClasses ,
             self :: use_strict_mode :: UseStrictMode ,
             self :: use_trim_start_end :: UseTrimStartEnd ,
