+++ crates/biome_js_analyze/tests/specs/nursery/noRestrictedImports/invalidPackagePrivateImports.options.json
@@ -1,0 +1,16 @@
+{
+	"$schema": "../../../../../../packages/@biomejs/biome/configuration_schema.json",
+	"linter": {
+		"enabled": true,
+		"rules": {
+			"nursery": {
+				"noRestrictedImports": {
+					"level": "error",
+					"options": {
+						"restrictPackagePrivate": "all"
+					}
+				}
+			}
+		}
+	}
+}
