+++ crates/biome_js_analyze/tests/specs/nursery/noRestrictedImports/validPackagePrivateImports.js
