+++ crates/biome_wasm/src/lib.rs
@@ -33,9 +33,6 @@ impl Workspace {
     #[allow(clippy::new_without_default)]
     pub fn new() -> Workspace {
         Workspace {
-            // Q: We can't use a real filesystem here, but is the memory
-            // filesystem the right choice? It might be, since I guess it
-            // will allow us to inject plugins if we wanted to.
             inner: workspace::server(Box::new(MemoryFileSystem::default())),
         }
     }
