+++ packages/@biomejs/biome/configuration_schema.json
@@ -2453,13 +2453,6 @@
 						{ "type": "null" }
 					]
 				},
-				"useImportRestrictions": {
-					"description": "Disallows package private imports.",
-					"anyOf": [
-						{ "$ref": "#/definitions/RuleConfiguration" },
-						{ "type": "null" }
-					]
-				},
 				"useNamedOperation": {
 					"description": "Enforce specifying the name of GraphQL operations.",
 					"anyOf": [
@@ -2664,6 +2657,21 @@
 			"type": "array",
 			"items": { "$ref": "#/definitions/OverridePattern" }
 		},
+		"PackagePrivateRestriction": {
+			"description": "Allowed values for the `restrictPackagePrivate` option.",
+			"oneOf": [
+				{
+					"description": "All exported symbols are assumed to be package private and may only be imported from sibling modules and their submodules.",
+					"type": "string",
+					"enum": ["all"]
+				},
+				{
+					"description": "No restrictions are applied on importing package private symbols.",
+					"type": "string",
+					"enum": ["none"]
+				}
+			]
+		},
 		"Performance": {
 			"description": "A list of rules that belong to this group",
 			"type": "object",
@@ -2758,6 +2766,11 @@
 					"description": "A list of names that should trigger the rule",
 					"type": "object",
 					"additionalProperties": { "type": "string" }
+				},
+				"restrictPackagePrivate": {
+					"description": "Whether to place restrictions on the importing of \"package private\" symbols.",
+					"default": "none",
+					"allOf": [{ "$ref": "#/definitions/PackagePrivateRestriction" }]
 				}
 			},
 			"additionalProperties": false
